requirements.txt
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
html5lib==1.0.1
idna==2.7
lxml==4.2.5
python-dateutil==2.7.5
requests==2.20.1
six==1.11.0
slackclient==1.3.0
urllib3==1.24.1
webencodings==0.5.1
websocket-client==0.54.0
